% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heading.R
\name{angAcc}
\alias{angAcc}
\title{Estimate instantaneous angular acceleration from a succession of coordinates}
\usage{
angAcc(x, y, t, geo = FALSE)
}
\arguments{
\item{x}{A vector of x (or longitude) coordinates corresponding to a single 
animal trajectory.}

\item{y}{A vector of y (or latitude) coordinates corresponding to a single 
animal trajectory.}

\item{t}{A vector of timestamps corresponding to a single animal trajectory.}

\item{geo}{A logical value indicating whether the locations are defined by 
geographic coordinates (pairs of longitude/latitude values). Default: FALSE.}
}
\value{
A vector of the same length as x, y and t corresponding to approximated 
 instantaneous angular acceleration along a [x, y] trajectory.
}
\description{
Given a set of cartesian coordinates representing an object's 
 trajectory, this function returns approximated instantaneous angular acceleration
 computed as the difference between successive angular speeds.
}
\examples{
# TODO

}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
\seealso{
\code{\link{heading}}, \code{\link{angAcc}}
}

