% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/heading.R
\name{angular_acceleration}
\alias{angular_acceleration}
\title{Compute instantaneous angular acceleration}
\usage{
angular_acceleration(s)
}
\arguments{
\item{s}{Vector of angular speeds}
}
\value{
This function returns a vector with the same length as s
corresponding to the instantaneous angular acceleration at each time step.
}
\description{
Given a series of angular speeds as computed by \code{\link{angular_speed}},
this function returns the angular acceleration as the difference between
successive angular speeds.
}
\examples{
# Generate and plot a random smooth trajectory
x <- filter(cumsum(rnorm(1001)), rep(1 / 10, 10))
y <- filter(cumsum(rnorm(1001)), rep(1 / 10, 10))
t <- 0:1000
plot(y ~ x, type = "l", asp = 1)

# Compute and plot distribution of angular accelerations
h <- heading(x, y)
as <- angular_speed(h, t)
aa <- angular_acceleration(as)
hist(aa)
}
\author{
Simon Garnier: \email{garnier@njit.edu}, \link[https://twitter.com/sjmgarnier]{@sjmgarnier}
}
\seealso{
\code{\link{heading}}, \code{\link{angular_speed}}
}

