% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/heading.R
\name{angular_speed}
\alias{angular_speed}
\title{Compute instantaneous angular speed}
\usage{
angular_speed(h, t)
}
\arguments{
\item{h}{Vector of headings}

\item{t}{Vector of time stamps}
}
\value{
This function returns a vector with the same length as h and t
corresponding to the instantaneous angular speed at each time step.
}
\description{
Given a series of headings as computed by \code{\link{heading}}, this
function returns the angular speed as the difference between successive
headings divided by the time between these successive headings.
}
\examples{
# Generate and plot a random smooth trajectory
x <- filter(cumsum(rnorm(1001)), rep(1 / 10, 10))
y <- filter(cumsum(rnorm(1001)), rep(1 / 10, 10))
t <- 0:1000
plot(y ~ x, type = "l", asp = 1)

# Compute and plot distribution of angular speeds
h <- heading(x, y)
as <- angular_speed(h, t)
hist(as)
}
\author{
Simon Garnier: \email{garnier@njit.edu}, \link[https://twitter.com/sjmgarnier]{@sjmgarnier}
}
\seealso{
\code{\link{heading}}, \code{\link{angular_acceleration}}
}

