% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centroid.R
\name{centroid}
\alias{centroid}
\title{Centroid of a set of coordinates}
\usage{
centroid(x, y, geo = FALSE)
}
\arguments{
\item{x}{A vector of x (or longitude) coordinates.}

\item{y}{A vector of y (or latitude) coordinates.}

\item{geo}{A logical value indicating whether the locations are defined by 
geographic coordinates (pairs of longitude/latitude values). Default: FALSE.}
}
\value{
A two-element vector corresponding to the location of the centroid.
}
\description{
This function computes the centroid of a set of x-y (or 
 longitude-latitude) coordinates.
}
\examples{
# TODO

}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
\seealso{
\code{\link{dist2centroid}}
}

