% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTraj.R
\name{completeTraj}
\alias{completeTraj}
\title{Add missing timestamps}
\usage{
completeTraj(traj, begin = NULL, end = NULL, step = NULL, geo = FALSE)
}
\arguments{
\item{traj}{A trajectory data table as produced by the \code{\link{makeTraj}}
function.}

\item{begin}{A \code{\link{POSIXct}}-formatted date-time representing the 
desired starting timestamp of the trajectory. If not set, it defaults to the
earliest timestamp in the \code{time} column of \code{traj}.}

\item{end}{A \code{\link{POSIXct}}-formatted date-time representing the 
desired final timestamp of the trajectory. If not set, it defaults to the
latest timestamp in the \code{time} column of \code{traj}.}

\item{step}{A \code{\link{difftime}} object representing the time between two 
consecutive locations of the trajectory. If not set, it is set to the most 
common time difference between successive locations in \code{traj}.}

\item{geo}{A logical value indicating whether the locations are defined by 
geographic coordinates (pairs of longitude/latitude values). Default: FALSE.}
}
\value{
A data table (from the \code{\link[data.table:data.table-package]{data.table}}
 package) with the same number of columns as traj.
}
\description{
Convenience function to add missing timestamps in a trajectory 
 data table.
}
\examples{
# TODO

}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
\seealso{
\code{\link{makeTraj}}, \code{\link{fixTraj}}
}

