% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centroid.R
\name{dist2centroid}
\alias{dist2centroid}
\title{Distance to the centroid of a set of coordinates}
\usage{
dist2centroid(x, y, geo = FALSE)
}
\arguments{
\item{x}{A vector of x (or longitude) coordinates.}

\item{y}{A vector of y (or latitude) coordinates.}

\item{geo}{A logical value indicating whether the locations are defined by 
geographic coordinates (pairs of longitude/latitude values). Default: FALSE.}
}
\value{
A vector of the same length as x and y corresponding to the 
 individual distance of each point to the centroid of the set.
}
\description{
Given a set of x-y (or longitude-latitude) coordinates, this 
 function computes their distance to the centroid of the set.
}
\examples{
# TODO

}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
