% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/distance_to_barycenter.R
\name{distance_to_barycenter}
\alias{distance_to_barycenter}
\title{Compute distance of individuals to the center of mass of the group}
\usage{
distance_to_barycenter(x, y)
}
\arguments{
\item{x}{Vector of x coordinates. Each value must represent a different individual.}

\item{y}{Vector of y coordinates. Each value must represent a different individual.}
}
\value{
This function returns a vector with the same length as x and y
corresponding to the individual distances to the group's center of mass.
}
\description{
Given the 2D coordinates of all the individuals in a group, this function
returns the individual distance to the center of mass (i.e. the average
location) of the group.
}
\examples{
# Create and plot random [x, y] coordinates
x <- rnorm(25)
y <- rnorm(25)
plot(y ~ x, asp = 1)

# Compute and plot center of mass
bx <- mean(x)
by <- mean(y)
points(by ~ bx, col = "red", pch = 19)

# Draw line between each point and the center of mass
mapply(
 function(x1, y1, x2, y2) {
   lines(c(y1, y2) ~ c(x1, x2), col = "red")
   },
 x, y,
 MoreArgs = list(x2 = bx, y2 = by)
)

# Compute distance to center of mass for each individual point and plot distribution
d <- distance_to_barycenter(x, y)
hist(d)
}
\author{
Simon Garnier: \email{garnier@njit.edu}, \link[https://twitter.com/sjmgarnier]{@sjmgarnier}
}

