% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{.cartesian_perimeter}
\alias{.cartesian_perimeter}
\title{Perimeter Of A Polygon In Cartesian Space}
\usage{
.cartesian_perimeter(x, y)
}
\arguments{
\item{x}{A vector of x coordinates.}

\item{y}{A vector of y coordinates.}
}
\description{
Given a set of Cartesian coordinates representing a polygon, 
 this function computes the perimeter of the polygon.
}
\seealso{
\code{\link{chull_perimeter}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
