% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{.cartesianPerimeter}
\alias{.cartesianPerimeter}
\title{Computes the perimeter of a polygon in cartesian space}
\usage{
.cartesianPerimeter(x, y)
}
\arguments{
\item{x}{A vector of x coordinates.}

\item{y}{A vector of y coordinates.}
}
\description{
Given a set of cartesian coordinates representing a polygon, 
 this function computes the perimeter of the polygon
}
\examples{
# TODO
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
