% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{.updateError}
\alias{.updateError}
\title{Update Error Description In Trajectory Tables}
\usage{
.updateError(error, update)
}
\arguments{
\item{error}{A character vector of error descriptions.}

\item{update}{A character string of the same length as \code{error} of the 
error descriptions to be appended to the current error descriptions.}
}
\description{
This is an internal utility function to update the description
 of errors in trajectory tables detected by the automated error detections 
 and correction functions of the package.
}
\examples{
# TODO
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
