% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R
\name{dplyr}
\alias{dplyr}
\alias{select.trackTable}
\alias{rename.trackTable}
\alias{filter.trackTable}
\alias{arrange.trackTable}
\alias{mutate.trackTable}
\alias{transmute.trackTable}
\alias{summarise.trackTable}
\alias{summarize.trackTable}
\alias{group_by.trackTable}
\alias{ungroup.trackTable}
\alias{sample_n.trackTable}
\alias{sample_frac.trackTable}
\alias{do.trackTable}
\title{Dplyr Methods For Track Tables}
\usage{
select.trackTable(.data, ...)

rename.trackTable(.data, ...)

filter.trackTable(.data, ...)

arrange.trackTable(.data, ...)

mutate.trackTable(.data, ...)

transmute.trackTable(.data, ...)

summarise.trackTable(.data, ...)

summarize.trackTable(.data, ...)

group_by.trackTable(.data, ...)

ungroup.trackTable(.data, ...)

sample_n.trackTable(.data, ...)

sample_frac.trackTable(.data, ...)

do.trackTable(.data, ...)
}
\arguments{
\item{.data}{A track table.}

\item{...}{Additional arguments to be passed to the corresponding 
\code{\link[dplyr]{dplyr}} method.}
}
\description{
\code{\link[dplyr]{dplyr}} methods for track tables objects.
}
\seealso{
\code{\link[dplyr]{select}}, \code{\link[dplyr]{rename}},
 \code{\link[dplyr]{filter}}, \code{\link[dplyr]{arrange}}, \code{\link[dplyr]{mutate}},
 \code{\link[dplyr]{transmute}}, \code{\link[dplyr]{summarise}}, \code{\link[dplyr]{summarize}},
 \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{ungroup}}, \code{\link[dplyr]{sample_n}},
 \code{\link[dplyr]{sample_frac}}, \code{\link[dplyr]{do}}
}
