% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{findLocNA}
\alias{findLocNA}
\title{Find NA locations in a trajectory table}
\usage{
findLocNA(traj)
}
\arguments{
\item{traj}{A trajectory data table as produced by the \code{\link{makeTraj}}
function.}
}
\value{
A vector corresponding to the row numbers of the NA locations in the 
 trajectory data table.
}
\description{
This function attempts to automatically detect NA locations in 
 trajectory tables.
}
\examples{
# TODO

}
\seealso{
\code{\link{makeTraj}}, \code{\link{fixLocNA}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
