% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{findLOCSEQ}
\alias{findLOCSEQ}
\title{Find inconsistent locations in a trajectory table}
\usage{
findLOCSEQ(traj, s = 15)
}
\arguments{
\item{traj}{A trajectory data table as produced by the \code{\link{makeTraj}}
function.}

\item{s}{The discrimination threshold of the outlier detection algorithm. 
Higher values correspond to less outliers.}
}
\value{
A vector corresponding to the row numbers of the inconsistent 
 locations in the trajectory data table.
}
\description{
This function attempts to automatically detect inconsistent 
 locations (for instance due to a writing error) in trajectory tables.
}
\examples{
# TODO

}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
\seealso{
\code{\link{makeTraj}}, \code{\link{fixLOCSEQ}}
}

