% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{findMISSING}
\alias{findMISSING}
\title{Find missing data in a trajectory table}
\usage{
findMISSING(traj, begin = NULL, end = NULL, step = NULL)
}
\arguments{
\item{traj}{A trajectory data table as produced by the \code{\link{makeTraj}}
function.}

\item{begin}{A full timestamp (date+time) in \code{\link{POSIXct}} format
corresponding to the beginning of the trajectory. If not set, it is set to 
the first timestamp of the trajectory table.}

\item{end}{A full timestamp (date+time) in \code{\link{POSIXct}} format
corresponding to the end of the trajectory. If not set, it is set to the 
last timestamp of the trajectory table.}

\item{step}{A \code{\link{difftime}} object representing the time between two 
consecutive locations of the trajectory. If not set, it is set to the most 
common time difference between successive locations in \code{traj}.}
}
\value{
A vector of the timestamps of the missing data points
}
\description{
This function attempts to automatically detect missing data (for 
 instance due to writing errors) in trajectory tables.
}
\examples{
# TODO

}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
\seealso{
\code{\link{makeTraj}}, \code{\link{fixMISSING}}
}

