% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{findTIMEDUP}
\alias{findTIMEDUP}
\title{Find duplicated timestamps in a trajectory table}
\usage{
findTIMEDUP(traj)
}
\arguments{
\item{traj}{A trajectory data table as produced by the \code{\link{makeTraj}}
function.}
}
\value{
A vector corresponding to the row numbers of the duplicated timestamps
 in the trajectory data table.
}
\description{
This function attempts to automatically detect duplicated 
 timestamps in trajectory tables.
}
\examples{
# TODO

}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
\seealso{
\code{\link{makeTraj}}, \code{\link{fixTIMEDUP}}
}

