% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{findTIMENA}
\alias{findTIMENA}
\title{Find NA timestamps in a trajectory table}
\usage{
findTIMENA(traj)
}
\arguments{
\item{traj}{A trajectory data table as produced by the \code{\link{makeTraj}}
function.}
}
\value{
A vector corresponding to the row numbers of the NA timestamps in the 
 trajectory data table.
}
\description{
This function attempts to automatically detect NA timestamps in 
 trajectory tables.
}
\examples{
# TODO

}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
\seealso{
\code{\link{makeTraj}}, \code{\link{fixTIMENA}}
}

