% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{findTIMESEQ}
\alias{findTIMESEQ}
\title{Find inconsistent timestamps in a trajectory table}
\usage{
findTIMESEQ(traj)
}
\arguments{
\item{traj}{A trajectory data table as produced by the \code{\link{makeTraj}}
function.}
}
\value{
A vector corresponding to the row numbers of the inconsistent timestamps
 in the trajectory data table.
}
\description{
This function attempts to automatically detect inconsistent 
 timestamps (for instance due to a writing error) in trajectory tables.
}
\examples{
# TODO

}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
\seealso{
\code{\link{makeTraj}}, \code{\link{fixTIMESEQ}}
}

