% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixes.R
\name{fixLocErr}
\alias{fixLocErr}
\title{Correct inconsistent locations in a trajectory table}
\usage{
fixLocErr(traj, s = 15, spline = FALSE)
}
\arguments{
\item{traj}{A trajectory data table as produced by the \code{\link{makeTraj}}
function.}

\item{s}{The discrmination threshold of the outlier detection algorithm. 
Higher values correspond to less outliers.}

\item{spline}{If \code{spline} is \code{TRUE}, inconsistent locations are 
estimated using spline interpolation. If \code{FALSE} (the default), a linear 
interpolation is used instead.}
}
\value{
A data table (from the \code{\link[data.table:data.table-package]{data.table}}
 package) with five columns:
 \itemize{
   \item{"id": }{the unique identifier of the trajectory.}
   \item{"x" or "lon": }{the x or longitude coordinates of the trajectory.}
   \item{"y" or "lat": }{the y or latitude coordinates of the trajectory.}
   \item{"time": }{the full timestamp (date+time) of each location in 
     \code{\link{POSIXct}} format.}
   \item{"error": }{the type of error detected and corrected (when possible) 
     by the function.}
 }
}
\description{
This function attempts to automatically detect and correct 
 inconsistent locations (for instance due to a writing error) in trajectory 
 tables.
}
\examples{
# TODO

}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
\seealso{
\code{\link{makeTraj}}, \code{\link{findLocErr}}
}

