% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixes.R
\name{fixLocNA}
\alias{fixLocNA}
\title{Correct NA locations in a trajectory table}
\usage{
fixLocNA(traj, spline = FALSE)
}
\arguments{
\item{traj}{A trajectory data table as produced by the \code{\link{makeTraj}}
function.}

\item{spline}{If \code{spline} is \code{TRUE}, NA locations are estimated 
using spline interpolation. If \code{FALSE} (the default), a linear 
interpolation is used instead.}
}
\value{
A trajectory data table as produced by the \code{\link{makeTraj}}
 function. If not present, an "error" column will be added that indicates 
 which observations were corrected.
}
\description{
This function attempts to automatically detect and correct 
 NA locations in trajectory tables.
}
\examples{
# TODO

}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
\seealso{
\code{\link{makeTraj}}, \code{\link{findLocNA}}
}

