% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixes.R
\name{fixMissing}
\alias{fixMissing}
\title{Interpolate missing data in a trajectory table}
\usage{
fixMissing(x, begin = NULL, end = NULL, step = NULL,
  spline = FALSE)
}
\arguments{
\item{x}{A trajectory data table as produced by the \code{\link{makeTraj}}
function.}

\item{begin}{A full timestamp (date+time) in \code{\link{POSIXct}} format
corresponding to the beginning of the trajectory. If not set, it is set to 
the first timestamp of the trajectory table.}

\item{end}{A full timestamp (date+time) in \code{\link{POSIXct}} format
corresponding to the end of the trajectory. If not set, it is set to the 
last timestamp of the trajectory table.}

\item{step}{A \code{\link{difftime}} object representing the time between two 
consecutive locations of the trajectory. If not set, it is set to the most 
common time difference between successive locations in \code{traj}.}

\item{spline}{If \code{spline} is \code{TRUE}, inconsistent locations are 
estimated using spline interpolation. If \code{FALSE} (the default), a linear 
interpolation is used instead.}
}
\value{
A trajectory data table as produced by the \code{\link{makeTraj}}
 function. If not present, an "error" column will be added that indicates 
 which observations were corrected.
}
\description{
This function attempts to automatically detect and correct 
 missing data (for instance due to writing errors) in trajectory tables.
}
\examples{
# TODO

}
\seealso{
\code{\link{makeTraj}}, \code{\link{findMissing}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
