% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixes.R
\name{fixTimeDup}
\alias{fixTimeDup}
\title{Correct duplicated timestamps in a trajectory table}
\usage{
fixTimeDup(traj, step = NULL)
}
\arguments{
\item{traj}{A trajectory data table as produced by the \code{\link{makeTraj}}
function.}

\item{step}{A \code{\link{difftime}} object representing the time between two 
consecutive locations of the trajectory. If not set, it is set to the most 
common time difference between successive locations in \code{traj}.}
}
\value{
A trajectory data table as produced by the \code{\link{makeTraj}}
 function. If not present, an "error" column will be added that indicates 
 which observations were corrected.
}
\description{
This function attempts to automatically detect and correct 
 duplicated timestamps in trajectory tables. If it cannot correct a duplicated
 timestamp, it replaces it with NA.
}
\examples{
# TODO

}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
\seealso{
\code{\link{makeTraj}}, \code{\link{findTimeDup}}
}

