% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTraj.R
\name{fixTraj}
\alias{fixTraj}
\title{Replace missing locations by interpolation}
\usage{
fixTraj(traj, geo = FALSE, spline = FALSE)
}
\arguments{
\item{traj}{A trajectory data table as produced by the \code{\link{makeTraj}}
function.}

\item{geo}{A logical value indicating whether the locations are defined by 
geographic coordinates (pairs of longitude/latitude values). Default: FALSE.}

\item{spline}{A logical value indicating whether the interpolated value should 
be calculated using cubic spline interpolation. Default: FALSE.}
}
\value{
A data table (from the \code{\link[data.table:data.table-package]{data.table}}
 package) with the same number of columns as traj.
}
\description{
Convenience function to deal with missing locations in a trajectory data 
 table. It replaces missing locations with interpolated locations using either
 linear or spline interpolation.
}
\examples{
# TODO

}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
\seealso{
\code{\link[stats:approx]{approx}}, \code{\link[stats:spline]{spline}},
 \code{\link{makeTraj}}, \code{\link{fillTraj}}
}

