% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/heading.R
\name{heading}
\alias{heading}
\title{Compute individual heading from a series of 2D coordinates}
\usage{
heading(x, y)
}
\arguments{
\item{x}{Vector of x coordinates}

\item{y}{Vector of y coordinates}
}
\value{
This function returns a vector with the same length as x and y
corresponding to the approximated headings along a [x, y] trajectory.
}
\description{
Given a series of 2D coordinates, this function returns approximated headings
computed as the vectors between successive points.
}
\examples{
# Generate and plot a random smooth trajectory
x <- filter(cumsum(rnorm(1001)), rep(1 / 10, 10))
y <- filter(cumsum(rnorm(1001)), rep(1 / 10, 10))
plot(y ~ x, type = "l", asp = 1)

# Compute and plot distribution of headings
h <- heading(x, y)
hist(h)
}
\author{
Simon Garnier: \email{garnier@njit.edu}, \link[https://twitter.com/sjmgarnier]{@sjmgarnier}
}
\seealso{
\code{\link{angular_speed}}, \code{\link{angular_acceleration}}
}

