% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackTable.R
\name{is.2D}
\alias{is.2D}
\title{Check if trajectory table is 2D or 3D}
\usage{
is.2D(x)
}
\arguments{
\item{x}{A trajectory data table as produced by the \code{\link{makeTraj}}
function.}
}
\description{
Trajectory tables produced by the \code{\link{makeTraj}} 
 function can have 2 or 3 dimensions. This function helps determine which is 
 being used in a particular table.
}
\examples{
# TODO

}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
