% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackTable.R
\name{is.trackTable}
\alias{is.trackTable}
\title{Check validity of trajectory table}
\usage{
is.trackTable(x)
}
\arguments{
\item{object}{An object to test.}
}
\description{
Test whether a variable contains a trajectory table as produced 
 by the \code{\link{makeTraj}} function.
}
\examples{
# TODO

}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
