% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/is_chull.R
\name{is_chull}
\alias{is_chull}
\title{Is point part of the convex hull of a set of points?}
\usage{
is_chull(x, y)
}
\arguments{
\item{x}{Vector of x coordinates}

\item{y}{Vector of y coordinates}
}
\value{
This function returns a logical vector with the same length as x and
y. TRUE indicates that the corresponding point is part of the convex hull of
the set.
}
\description{
Given a set of 2D coordinates, this function return whether each points
belongs to the convex hull (or envelope) of the set.
}
\examples{
# Create and plot random [x, y] coordinates
x <- rnorm(25)
y <- rnorm(25)
plot(y ~ x, asp = 1)

# Compute which point is part of the convex hull
is_ch <- is_chull(x, y)
ch <- chull(x, y)

# Plot convex hull
points(y[is_ch] ~ x[is_ch], col = "red", pch = 19)
lines(c(y[ch], y[ch[1]]) ~ c(x[ch], x[ch[1]]), col = "red")
}
\author{
Simon Garnier: \email{garnier@njit.edu}, \link[https://twitter.com/sjmgarnier]{@sjmgarnier}
}
\seealso{
\code{\link[grDevices]{linear_acceleration}}
}

