% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isChull.R
\name{isChull}
\alias{isChull}
\title{Is a point part of the convex hull of a set of points?}
\usage{
isChull(x, y)
}
\arguments{
\item{x}{A vector of x (or longitude) coordinates.}

\item{y}{A vector of y (or latitude) coordinates.}
}
\value{
A logical vector of the same length as x and y. TRUE indicates that 
the corresponding point is part of the convex hull of the set.
}
\description{
Given a set of cartesian coordinates representing, this function 
 determines which of these points belongs to the convex hull (or envelope) of 
 the set.
}
\examples{
# TODO

}
\author{
Simon Garnier, \email{garnier@njit.edu}
}

