% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{isGeo}
\alias{isGeo}
\title{Check if trajectory table is using geographic coordinates}
\usage{
isGeo(traj)
}
\arguments{
\item{traj}{A trajectory data table as produced by the \code{\link{makeTraj}}
function.}
}
\description{
Trajectory tables produced by the \code{\link{makeTraj}} 
 function can use a cartesian (x, y) or a geographic (latitude, longitude) 
 coordinate system. This function helps determine which is being used in a 
 particular table.
}
\examples{
# TODO

}
\author{
Simon Garnier, \email{garnier@njit.edu}
}

