% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{isTraj}
\alias{isTraj}
\title{Check validity of trajectory table}
\usage{
isTraj(traj)
}
\arguments{
\item{traj}{A variable to test.}
}
\description{
Test whether a variable contains a trajectory table as produced 
 by the \code{\link{makeTraj}} function.
}
\examples{
# TODO

}
\author{
Simon Garnier, \email{garnier@njit.edu}
}

