% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speed.R
\name{linDist}
\alias{linDist}
\title{Linear distances along a trajectory}
\usage{
linDist(x, y, geo = FALSE)
}
\arguments{
\item{x}{A vector of x (or longitude) coordinates corresponding to a single 
animal trajectory.}

\item{y}{A vector of y (or latitude) coordinates corresponding to a single 
animal trajectory.}

\item{geo}{A logical value indicating whether the locations are defined by 
geographic coordinates (pairs of longitude/latitude values). Default: FALSE.}
}
\value{
A vector of the same length as x and y corresponding to the linear 
 distances between each pair of successive locations along the trajectory.
}
\description{
Given a set of cartesian coordinates representing an object's 
 trajectory, this function computes the linear distances between each pair of 
 successive locations along the trajectory.
}
\examples{
# TODO

}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
\seealso{
\code{\link{linSpeed}}, \code{\link{linAcc}}
}

