% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/speed.R
\name{linear_speed}
\alias{linear_speed}
\title{Compute instantaneous linear speed}
\usage{
linear_speed(x, y, t)
}
\description{
Given a series of spatio-temporal 2D coordinates, this function returns the
instantaneous linear speed as the distance between successive positions
divided by the time between these successive positions.
}

