% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTraj.R
\name{makeTraj}
\alias{makeTraj}
\title{Build a trajectory table}
\usage{
makeTraj(x, y, id = NULL, date = NULL, time = NULL, tz = "UTC",
  fps = NULL, geo = FALSE)
}
\arguments{
\item{x}{A vector of x (or longitude) coordinates corresponding to a single 
animal trajectory.}

\item{y}{A vector of y (or latitude) coordinates corresponding to a single 
animal trajectory.}

\item{id}{A unique identifier for the trajectory. It can be a number or a 
character string. If not set, it defaults to zero.}

\item{date}{A character string corresponding to a date in year-month-day 
format. Alternatively, a vector of character strings of the same length as x
and y (for instance if the trajectory spans multiple days). If not set, it 
default to January 1st, 1970.}

\item{time}{A vector of character strings corresponding to the timestamps of 
each location of the trajectory, in hours-minutes-seconds format. If not set,
the first timestamp defaults to 0 hours, 0 minutes, 1 second. The following 
timestamps are either in increment of 1 second or 1 second / fps.}

\item{tz}{A character string identifying the timezone of the timestamps of 
the trajectory. Default: "UTC".}

\item{fps}{A single numeric value corresponding to the sampling rate of the 
trajectory.}

\item{geo}{A logical value indicating whether the locations are defined by 
geographic coordinates (pairs of longitude/latitude values). Default: FALSE.}
}
\value{
A data table (from the \code{\link[data.table:data.table-package]{data.table}}
 package) with four columns:
 \itemize{
   \item{"id": }{the unique identifier of the trajectory.}
   \item{"x" or "lon": }{the x or longitude coordinates of the trajectory.}
   \item{"y" or "lat": }{the y or latitude coordinates of the trajectory.}
   \item{"time": }{the full timestamp (date+time) of each location in 
     \code{\link{POSIXct}} format.}
 }
}
\description{
Given a set of cartesian coordinates representing an object's trajectory, 
 this function builds a trajectory table that can then be used to work with 
 the functions provided by the \code{\link[swaRm:swaRm-package]{swaRm}}
 package.
}
\examples{
# TODO

}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
\seealso{
\code{\link{fillTraj}}, \code{\link{fixTraj}}
}

