% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{Mode}
\alias{Mode}
\title{Compute the mode(s) of a discrete distribution}
\usage{
Mode(x, na.rm = TRUE)
}
\arguments{
\item{x}{A vector or matrix of discrete values.}
}
\value{
A vector of values corresponding to the mode(s) of x.
}
\description{

}
\examples{
# TODO

}
\author{
Simon Garnier, \email{garnier@njit.edu}
}

