% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/nearest_neighbor.R
\name{nearest_neighbor}
\alias{nearest_neighbor}
\title{Find nearest neihgbor}
\usage{
nearest_neighbor(x, y)
}
\arguments{
\item{x}{Vector of x coordinates}

\item{y}{Vector of y coordinates}
}
\value{
This function returns a data frame with 3 columns and with the same
number of rows as the lengths of x and y. Each row corresponds to an
individual. Column \code{id} corresponds to the identity of each individual.
Column \code{nn} corresponds to the identity of the nearest neighbor of each
individual. Column \code{nnd} corresponds to the distance between each
individual and its nearest neighbor.
}
\description{
Find the nearest neighbor of each point in a set of 2D points, and compute
the nearest neighbor distances.
}
\examples{
# Generate and plot random points
x <- rnorm(25)
y <- rnorm(25)
plot(y ~ x, asp = 1)

# Compute nearest neighbors
neighbors <- nearest_neighbor(x, y)

# Plot nearest neighbor network (arrow points from the focal individual to its
# nearest neighbor)
apply(neighbors, 1,
      function(n, x, y) {
        arrows(x[n[1]], y[n[1]], x[n[2]], y[n[2]], length = 0.1, col = "red")
      },
      x = x, y = y
)

# Plot distribution of nearest neighbor distances
hist(neighbors$nnd)
}
\author{
Simon Garnier: \email{garnier@njit.edu}, \link[https://twitter.com/sjmgarnier]{@sjmgarnier}
}

