% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{nsd}
\alias{nsd}
\title{Net squared displacement}
\usage{
nsd(x, y, geo = FALSE)
}
\arguments{
\item{x}{A vector of x (or longitude) coordinates corresponding to a single 
animal trajectory.}

\item{y}{A vector of y (or latitude) coordinates corresponding to a single 
animal trajectory.}

\item{geo}{A logical value indicating whether the locations are defined by 
geographic coordinates (pairs of longitude/latitude values). Default: FALSE.}
}
\value{
A vector of the same length as x and y corresponding to the net 
 squared distances between each location and the first location of the 
 trajectory.
}
\description{
Given a set of cartesian coordinates representing an object's 
 trajectory, this function computes the net squared displacement of this 
 object, that is the squared distances between each location and the first
 location of the trajectory
}
\examples{
# TODO

}
\seealso{
\code{\link{linDist}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
