% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn.R
\name{pdist}
\alias{pdist}
\title{Pairwise distance matrix}
\usage{
pdist(x, y, geo = FALSE)
}
\arguments{
\item{x}{A vector of x (or longitude) coordinates.}

\item{y}{A vector of y (or latitude) coordinates.}

\item{geo}{A logical value indicating whether the locations are defined by 
geographic coordinates (pairs of longitude/latitude values). Default: FALSE.}
}
\value{
A length(x) * length(y) square matrix representing pairwise distances
 between all the objects.
}
\description{
Given a set of cartesian coordinates representing the locations 
 of different objects, this function computes the distances between each 
 possible pair of objects.
}
\examples{
# TODO

}
\seealso{
\code{\link{nn}}, \code{\link{nnd}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
