% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orders.R
\name{pol_order}
\alias{pol_order}
\alias{polOrder}
\title{Polarization Order Parameter}
\usage{
pol_order(h)

polOrder(h)
}
\arguments{
\item{h}{A vector of headings (in radians).}
}
\value{
A single value between 0 and 1 corresponding to the polarization 
 order parameter of the group.
}
\description{
Given a set of headings, this function returns the polarization 
 order of the set.
}
\examples{
h <- runif(25, 0, 2 * pi)
pol_order(h)

}
\seealso{
\code{\link{rot_order}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
