% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orders.R
\name{polOrder}
\alias{polOrder}
\title{Polarization order parameter}
\usage{
polOrder(h)
}
\arguments{
\item{h}{A vector of headings (in radians).}
}
\value{
A single value between 0 and 1 corresponding to the polarization 
order parameter of the group.
}
\description{
Given the headings of objects in a group, this function returns 
the polarization order of the group.
}
\examples{
# TODO

}
\seealso{
\code{\link{rotOrder}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
