% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rot_order.R
\name{rot_order}
\alias{rot_order}
\title{Compute rotational order parameter}
\usage{
rot_order(h, x, y)
}
\arguments{
\item{h}{Vector of headings}

\item{x}{Vector of x coordinates}

\item{y}{Vector of y coordinates}
}
\value{
This function returns a single value between 0 and 1 corresponding to
the rotational order parameter of the group.
}
\description{
Given the headings and positions of individuals in a group, this function
returns the rotational order of the group.
}
\examples{
# Compute and plot polarization order parameter for groups of 100 individuals
# with headings varying from all equals (noise = 0) to random (noise = pi)
noise <- seq(0, pi, pi/64)

order <- sapply(noise,
                function(n) {
                  pol_order(runif(100, min = pi - n, max = pi + n))
                }
)

plot(order ~ noise, ylim = c(0, 1))
}
\author{
Simon Garnier: \email{garnier@njit.edu}, \link[https://twitter.com/sjmgarnier]{@sjmgarnier}
}
\seealso{
\code{\link{pol_order}}
}

