% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orders.R
\name{rotOrder}
\alias{rotOrder}
\title{Rotational order parameter}
\usage{
rotOrder(h, x, y)
}
\arguments{
\item{h}{A vector of headings (in radians).}

\item{x}{A vector of x (or longitude) coordinates.}

\item{y}{A vector of y (or latitude) coordinates.}
}
\value{
A single value between 0 and 1 corresponding to the rotational 
order parameter of the group.
}
\description{
Given the headings and cartesian coordinates of objects in a 
 group, this function returns the rotational order of the group.
}
\examples{
# TODO

}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
\seealso{
\code{\link{polOrder}}
}

