% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.R
\name{stretch}
\alias{stretch}
\title{Approximated stretching direction of a set of points}
\usage{
stretch(x, y)
}
\arguments{
\item{x}{A vector of x coordinates.}

\item{y}{A vector of y coordinates.}
}
\value{
A single numeric value corresponding to the angle (in radians) of the 
 main axis of the bivariate 95% confidence ellipse of the set.
}
\description{
Given a set of cartesian coordinates, this function approximates 
 the stretching direction of the set by calculating the angle of the main 
 axis of the bivariate 95% confidence ellipse of the set.
}
\examples{
# TODO

}
\author{
Simon Garnier, \email{garnier@njit.edu}
}

