% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swaRm.R
\docType{package}
\name{swaRm-package}
\alias{swaRm}
\alias{swaRm-package}
\title{Analysis tools for collective animal movement data}
\description{
This package contains functions to facilitate and automate the analysis of 
collective animal movement data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://swarm-lab.github.io/swaRm/}
  \item \url{https://github.com/swarm-lab/swaRm}
  \item Report bugs at \url{https://github.com/swarm-lab/swaRm/issues}
}

}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
