% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackTable.R
\name{track}
\alias{track}
\title{Build Track}
\usage{
track(id = NULL, x, y, z = NULL, t, t0 = NULL, units = "secs",
  geo = FALSE, tz = NULL)
}
\arguments{
\item{id}{A unique identifier for the trajectory. It can be a number or a 
character string. If not set, it defaults to 0. Multiple track identifiers 
can be passed to the function, in which case \code{id} should be a vector of 
the same length as \code{x}.}

\item{x, y, z}{Vectors of coordinates. Note that \code{z} is not required. If 
\code{geo=TRUE}, then \code{x} will correspond to longitudes, \code{y} to 
latitudes and \code{z} to altitudes.}

\item{t}{A numeric vector of the time elapsed since the beginning of the 
tracking. The time units are specified in \code{units}.}

\item{t0}{The time at which tracking began. It can be any object that can be 
converted to a date-time by \code{\link[lubridate]{as_datetime}}. If not 
specified, it defaults to the date-time returned by \code{\link[lubridate]{now}}.}

\item{units}{A character string indicating the units in which \code{t} as been 
recorded (default: "secs"). See \code{\link{units}} for more details.}

\item{geo}{A logical value indicating whether the locations are defined by 
geographic coordinates (default: FALSE).}

\item{tz}{An optional time zone specification for \code{t0}. If not specified,
it defaults to the system's time zone. See \code{\link{timezone}} for more \
details.}
}
\value{
A trajectory table.
}
\description{
Given a set of coordinates representing an object's trajectory, 
 this function builds a trajectory table that can then be used to work with 
 the functions provided by the \code{\link[swaRm:swaRm-package]{swaRm}} package.
}
\examples{
# TODO

}
\seealso{
\code{\link[lubridate]{as_datetime}}, \code{\link[lubridate]{now}}, 
 \code{\link{units}}, \code{\link{timezone}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
