% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackTable.R
\name{trackTable}
\alias{trackTable}
\title{Track table class}
\usage{
trackTable(id, t, x, y, z = NULL, geo = FALSE)
}
\arguments{
\item{id}{A vector of identifiers. Each identifier should correspond to the 
coordinates of a single trajectory. Identifiers can be numbers or character
strings.}

\item{t}{A vector of standardized date-time values.}

\item{x}{A vector of x (or longitude) coordinates.}

\item{y}{A vector of y (or latitude) coordinates.}

\item{z}{A vector of z (or altitude) coordinates.}

\item{geo}{A logical value indicating whether the locations are defined by 
geographic coordinates (longitude/latitude/altitude values) or projected 
coordinates (x/y/z values in meters). Default: FALSE.}
}
\value{
A trajectory table with 4 or 5 columns:
 \itemize{
   \item{"id": }{the unique identifier of the trajectory.}
   \item{"time": }{the full timestamp (date+time) of each location in 
     \code{\link{POSIXct}} format.}
   \item{"x" or "lon": }{the x or longitude coordinates of the trajectories.}
   \item{"y" or "lat": }{the y or latitude coordinates of the trajectories.}
   \item{"z" or "alt": }{the z or altitude coordinates of the trajectories.}
 }
}
\description{
A track table is a regular \code{\link[data.table:data.table-package]{data.table}}
 enhanced to be used by the functions in \code{\link[swaRm:swaRm-package]{swaRm}}.
 It is a standardized way to store geographic or projected, single or group, 
 2D or 3D trajectories.
}
\examples{
# TODO

}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
\seealso{
\code{\link{makeTraj}}
}

