% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{.ellipse}
\alias{.ellipse}
\title{Computes confidence ellipse of a bivariate set of points}
\usage{
.ellipse(x, y, level = 0.95)
}
\arguments{
\item{x}{A vector of x coordinates.}

\item{y}{A vector of y coordinates.}

\item{level}{The confidence level of the ellipse.}
}
\description{
This function computes the confidence ellipse of a set of 
 bivariate coordinates.
}
\examples{
# TODO
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
