% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{findTimeDup}
\alias{findTimeDup}
\title{Find duplicated timestamps in a trajectory table}
\usage{
findTimeDup(x)
}
\arguments{
\item{x}{A trajectory data table as produced by the \code{\link{makeTraj}}
function.}
}
\value{
A vector corresponding to the row numbers of the duplicated timestamps
 in the trajectory data table.
}
\description{
This function attempts to automatically detect duplicated 
 timestamps in trajectory tables.
}
\note{
If a timestamps is missing where the duplicated timestamp is, this 
 function will replace the duplicated timestamp with the missing one. 
 Otherwise, the duplicated timestamp will be replaced by NA.
}
\examples{
# TODO

}
\seealso{
\code{\link{makeTraj}}, \code{\link{fixTimeDup}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
