% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixes.R
\name{fixLocErr}
\alias{fixLocErr}
\title{Correct inconsistent locations in a trajectory table}
\usage{
fixLocErr(traj, s = 15, spline = FALSE)
}
\arguments{
\item{traj}{A trajectory data table as produced by the \code{\link{makeTraj}}
function.}

\item{s}{The discrmination threshold of the outlier detection algorithm. 
Higher values correspond to less outliers.}

\item{spline}{If \code{spline} is \code{TRUE}, inconsistent locations are 
estimated using spline interpolation. If \code{FALSE} (the default), a linear 
interpolation is used instead.}
}
\value{
A trajectory data table as produced by the \code{\link{makeTraj}}
 function. If not present, an "error" column will be added that indicates 
 which observations were corrected.
}
\description{
This function attempts to automatically detect and correct 
 inconsistent locations (for instance due to a writing error) in trajectory 
 tables.
}
\examples{
# TODO

}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
\seealso{
\code{\link{makeTraj}}, \code{\link{findLocErr}}
}

