% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heading.R
\name{heading}
\alias{heading}
\title{Estimate headings from a succession of coordinates}
\usage{
heading(x, y, geo = FALSE)
}
\arguments{
\item{x}{A vector of x (or longitude) coordinates corresponding to a single 
animal trajectory.}

\item{y}{A vector of y (or latitude) coordinates corresponding to a single 
animal trajectory.}

\item{geo}{A logical value indicating whether the locations are defined by 
geographic coordinates (pairs of longitude/latitude values). Default: FALSE.}
}
\value{
A vector of the same length as x and y corresponding to approximated 
 headings along a [x, y] trajectory.
}
\description{
Given a set of cartesian coordinates representing an object's 
 trajectory, this function returns approximated headings computed as the 
 direction of vectors between successive points.
}
\examples{
# TODO

}
\seealso{
\code{\link{angSpeed}}, \code{\link{angAcc}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
