% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isChull.R
\name{isChull}
\alias{isChull}
\title{Is a point part of the convex hull of a set of points?}
\usage{
isChull(x, y)
}
\arguments{
\item{x}{A vector of x (or longitude) coordinates.}

\item{y}{A vector of y (or latitude) coordinates.}
}
\value{
A numerical vector of the same length as \code{x} and \code{y}. 
 \code{0} indicates that the corresponding point is not part of the convex 
 hull of the set. Values \code{>0} indicates that the corresponding point is 
 part of the convex hull, and each value corresponds to the order of the 
 point along the convex hull polygon.
}
\description{
Given a set of cartesian coordinates, this function determines 
 which points belongs to the convex hull (or envelope) of the set.
}
\examples{
# TODO

}
\author{
Simon Garnier, \email{garnier@njit.edu}
}

