% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/speed.R
\name{linear_acceleration}
\alias{linear_acceleration}
\title{Compute instantaneous linear acceleration}
\usage{
linear_acceleration(s)
}
\description{
Given a series of instantaneous linear speeds as computed by \link[swarmR]{linear_speed},
this function returns the instantaneous linear acceleration as the difference
between successive instantaneous linear speeds.
}

