% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn.R
\name{nn}
\alias{nn}
\title{Nearest neihgbor}
\usage{
nn(x, y, geo = FALSE)
}
\arguments{
\item{x}{A vector of x (or longitude) coordinates.}

\item{y}{A vector of y (or latitude) coordinates.}

\item{geo}{A logical value indicating whether the locations are defined by 
geographic coordinates (pairs of longitude/latitude values). Default: FALSE.}
}
\value{
A vector of the same length as x and y representing the identity of 
 the nearest neighboring object to each object.
}
\description{
Given a set of cartesian coordinates representing the locations 
 of different objects, this function determines the nearest neighboring object 
 of each object.
}
\examples{
# TODO

}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
\seealso{
\code{\link{nn}}, \code{\link{nnd}}
}

