% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn.R
\name{nnd}
\alias{nnd}
\title{Nearest neihgbor distance}
\usage{
nnd(x, y, geo = FALSE)
}
\arguments{
\item{x}{A vector of x (or longitude) coordinates.}

\item{y}{A vector of y (or latitude) coordinates.}

\item{geo}{A logical value indicating whether the locations are defined by 
geographic coordinates (pairs of longitude/latitude values). Default: FALSE.}
}
\value{
A vector of the same length as x and y representing the distance to 
 the nearest neighboring object for each object.
}
\description{
Given a set of cartesian coordinates representing the locations 
 of different objects, this function computes the distance to the nearest 
 neighboring object for each object.
}
\examples{
# TODO

}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
\seealso{
\code{\link{nn}}, \code{\link{nnd}}
}

