% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swaRm.R
\docType{package}
\name{swaRm-package}
\alias{swaRm}
\alias{swaRm-package}
\title{Analysis tools for collective animal movement data}
\description{
This package contains functions to facilitate and automate the analysis of 
collective animal movement data.
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}

